correct references to models after package renames
