+++ org.openjdk.jmc.console.ext.agent/src/main/java/org/openjdk/jmc/console/ext/agent/manager/model/IMethodReturnValue.java
@@ -33,5 +33,5 @@
  */
 package org.openjdk.jmc.console.ext.agent.manager.model;
 
-public interface IMethodReturnValue extends INamedCapturedValue {
+public interface IMethodReturnValue extends ICapturedValue {
 }
